\name{SDAMS-package}
\docType{package}
\alias{SDAMS-package}

\title{
 SDAMS package for differential abundance analysis of Metabolomics and
 Proteomics data from mass spectrometry
}
\description{
  SDAMS is an R package for differential abundance analysis of Metabolomics and
  Proteomics data, and the main function for differential abundance analysis
  is \code{\link{SDA}}. See the examples at \code{\link{SDA}} for
  basic analysis steps. SDAMS considers a two-part model, a logistic regression
  for the zero proportion and a semi-parametric log-linear model for the
  non-zero values
}

\author{
  Yuntong Li <yuntong.li@uky.edu>,
  Chi Wang <chi.wang@uky.edu>,
  Li Chen <lichenuky@uky.edu>
}

\keyword{package}
